requirements.txt
requests
tqdm
six
python-dateutil
tzlocal
